reports/requirements.txt
# Core Packages
python>=3.8
pandas>=1.3.0
numpy>=1.21.0

# API and Data Handling
requests>=2.26.0
python-dotenv>=0.19.0

# Data Visualization
matplotlib>=3.4.0
seaborn>=0.11.0
plotly>=5.3.0

# Data Processing
python-dateutil>=2.8.0
tqdm>=4.62.0  # For progress bars

# Jupyter Support (optional)
jupyter>=1.0.0
ipykernel>=6.0.0

# Documentation
mkdocs>=1.2.0  # For documentation generation
mkdocs-material>=7.3.0

# Testing
pytest>=6.2.0
pytest-cov>=2.12.0

# Code Quality
black>=21.0
flake8>=3.9.0
isort>=5.9.0